Add Max Metric to Webreport
Max Metric is now included in Webreport DTO.
In Webreport the user can switch between between multiple metrics:
If JPlag was run with a metric other than the max metric, e.g. avg metric,
the user will be able to rank comparisons in the WebReport by the
max metric additionally,
If JPlag was run with the max metric, this is the only metric available.